refactor(cluster): extract tag/source aggregation into ClusterTagAggregator (#133)
NewsClusterQueryService grew to 585 lines after PR-4a/4b added feedbackType and
suggestedQuestions. sources/stocks/topics aggregation was duplicated between
feed and detail paths, and the class mixed query orchestration with aggregation.

Component extraction:
- ClusterTagAggregator (@Component) owns sources/stocks/topics aggregation
  · aggregateSources, aggregateStocks, aggregateMarketTags (feed: bulk map)
  · aggregateStocksForCluster, aggregateMarketTagsForCluster (detail: single)
  · aggregateDetailSources (detail: section-prioritized, stable-sorted)
- NewsClusterQueryService delegates aggregation (9 deps -> 8; tag repo dropped)

Shared DTO package:
- New package domain.news.cluster.dto for SourceInfo, StockInfo, ArticleSourceInfo
  (previously nested public records inside NewsClusterQueryService)
- Follows CLAUDE.md 'domain common DTO' convention; decouples web DTO imports
  from service internals
- ClusterFeedResult / ClusterDetailResult / ClusterFeedItem / SectionDetail
  stay inside NewsClusterQueryService (service-result records)

Detail path cleanup:
- Single-cluster stocks/topics now use the *ForCluster overloads; drops the
  Map.of(clusterId, articleIds) wrapping and the redundant getOrDefault lookup
- aggregateDetailSources rewritten as a single stable-sorted stream
  (prioritized ? 0 : 1), drops the two-pass for-loop

Size: NewsClusterQueryService 585 -> 453 (-132 lines, 22% reduction);
ClusterTagAggregator 196 lines.
